requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ImageHash==4.3.1
numpy==2.0.1
openai==1.37.2
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
requests==2.32.3
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
